// src/types/post.ts
export interface Post {
  userId: number;
  id: number;
  title: string;
  body: string;
}
export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
  };
}

export interface PostComment {
  postId: number;
  id: number;
  name: string;
  email: string;
  body: string;
}